backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Celery
celery>=5.3.6
redis>=5.0.1

# APIs
openai>=1.12.0
elevenlabs>=1.50.3
httpx>=0.26.0

# PPTX processing
python-pptx>=0.6.23
Pillow>=10.2.0

# Utils
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
aiofiles>=23.2.1

# Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
aiosqlite>=0.19.0

